Частичная реализация Proxy системы

+++ ProxyHostServer/proxy-host-server/Data/GameServer/Info/GameServerInfo.cs
@@ -1,4 +1,5 @@
 ﻿using phs.Data.GameServer.Network;
+using phs.Data.NodeHostServer.Info;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -7,7 +8,7 @@ namespace phs.Data.GameServer.Info
 {
     public class GameServerInfo
     {
-        public short Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Текущий клиент для передачи данных
